backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0 # Use [standard] for performance
sqlalchemy>=2.0.0
alembic>=1.11.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
psycopg2-binary>=2.9.6 # For PostgreSQL connection
prophet>=1.1.4 # For forecasting
scikit-learn>=1.3.0 # For potential ML tasks
pandas>=2.0.0 # Required by Prophet
requests>=2.31.0 # For Dynamics BC API calls
apscheduler>=3.10.1 # For background tasks (optional)
python-dotenv>=1.0.0 # For loading .env files

